Add failover sample for Lettuce
* change package name from redislabs to redis
* refactor code
* update spring boot, jedis and lettuce versions
* add lettuce failover sample
